Update to code to combine into 1 sample
